Added Player class. Changed dice outcomes to bytes. Totals of outcomes remain as ints. Changed bet amounts, table min and max to uints. Added dictionary to bet factory to store payout ratio information (looked up prior to bet creation). Bet amounts consider table minimum.

+++ CrapsLibrary/Die.cs
@@ -2,18 +2,18 @@
 {
     public class Die
     {
-        private int numSides;
+        private byte numSides;
 
         static Random Generator = new Random();
         
-        public Die(int numSides)
+        public Die(byte numSides)
         {
             this.numSides = numSides;
         }
 
-        public int NewOutcome()
+        public byte NewOutcome()
         {
-            return Generator.Next(1, this.numSides + 1);
+            return (byte)Generator.Next(1, this.numSides + 1);
         }
     }
 }
