+++ CrapsLibrary/Scoreboard.cs
@@ -2,11 +2,11 @@
 {
     public class Scoreboard // a publisher
     {
-        public List<int> die01Rolls;
+        public List<byte> die01Rolls;
 
-        public List<int> die02Rolls;
+        public List<byte> die02Rolls;
 
-        public delegate void OnDiceRolled(int outcome01, int outcome02);
+        public delegate void OnDiceRolled(byte outcome01, byte outcome02);
 
         public List<OnDiceRolled> ListOfBetsToEvaluate = new List<OnDiceRolled>();
 
@@ -14,8 +14,8 @@
 
         public Scoreboard()
         {
-            this.die01Rolls = new List<int>();
-            this.die02Rolls = new List<int>();
+            this.die01Rolls = new List<byte>();
+            this.die02Rolls = new List<byte>();
         }
 
         public void NewSubscriber(OnDiceRolled newBet)
